Revert "Refactor Slack API operations by introducing SlackApiService"
This reverts commit a53e5779c52f66eec62fc68c844d3bbf59f48c77.
