+++ app/slack_api_ops.py
@@ -1,0 +1,74 @@
+from typing import Optional
+
+from slack_sdk.web import SlackResponse, WebClient
+
+# ----------------------------
+# General operations in a channel
+# ----------------------------
+
+
+def find_parent_message(
+    client: WebClient, channel_id: Optional[str], thread_ts: Optional[str]
+) -> Optional[dict]:
+    if channel_id is None or thread_ts is None:
+        return None
+
+    messages: list[dict] = client.conversations_history(
+        channel=channel_id,
+        latest=thread_ts,
+        limit=1,
+        inclusive=True,
+    ).get("messages", [])
+
+    return messages[0] if messages else None
+
+
+def is_this_app_mentioned(bot_user_id: Optional[str], parent_message: dict) -> bool:
+    parent_message_text = parent_message.get("text", "")
+    return f"<@{bot_user_id}>" in parent_message_text
+
+
+# ----------------------------
+# WIP reply message stuff
+# ----------------------------
+
+
+def post_wip_message(
+    *,
+    client: WebClient,
+    channel: str,
+    thread_ts: Optional[str],
+    loading_text: str,
+    messages: list[dict],
+    user: str,
+) -> SlackResponse:
+    system_messages = [msg for msg in messages if msg["role"] == "system"]
+    return client.chat_postMessage(
+        channel=channel,
+        thread_ts=thread_ts,
+        text=loading_text,
+        metadata={
+            "event_type": "litellm-convo",
+            "event_payload": {"messages": system_messages, "user": user},
+        },
+    )
+
+
+def update_wip_message(
+    client: WebClient,
+    channel: str,
+    ts: str,
+    text: str,
+    messages: list[dict],
+    user: str,
+) -> SlackResponse:
+    system_messages = [msg for msg in messages if msg["role"] == "system"]
+    return client.chat_update(
+        channel=channel,
+        ts=ts,
+        text=text,
+        metadata={
+            "event_type": "litellm-convo",
+            "event_payload": {"messages": system_messages, "user": user},
+        },
+    )
